[CAL][4] - Implement MsgsBetweenSeqNums() function in accessor (#987)
